pyproject.toml
[project]
name = "mcpclient"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "mcp>=1.9.0",
    "ollama>=0.4.8",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
    "textual>=3.2.0",
    "tiktoken>=0.9.0",
    "transformers>=4.51.3",
]
